+++ rmit-combined/recipes/configure.rb
@@ -1,8 +1,10 @@
 cwb_benchmark_suite 'rmit-benchmark-suite'
 
 # Implementations provided under `files/default/benchmark_name.rb`
-cwb_benchmark 'sysbench-cpu'
-cwb_benchmark 'sysbench-memory'
-cwb_benchmark 'sysbench-threads'
-cwb_benchmark 'sysbench-mutex'
-cwb_benchmark 'sysbench-fileio'
+# cwb_benchmark 'sysbench-cpu'
+# cwb_benchmark 'sysbench-memory'
+# cwb_benchmark 'sysbench-threads'
+# cwb_benchmark 'sysbench-mutex'
+# cwb_benchmark 'sysbench-fileio'
+
+cwb_benchmark 'fio'